chore: move applyFileWithConfirm to services
